Create re-usable component and clean up routing controller
